+++ gcpy/util.py
@@ -660,35 +660,37 @@ def get_diff_of_diffs(
     ref = ref[varlist]
     dev = dev[varlist]
 
-    # ------------------------------------------------------------------
-    # Compute absolute (Dev-Ref) and fractional (Dev/Ref) differences
-    # ------------------------------------------------------------------
+    # Preseve xarray attributes
     with xr.set_options(keep_attrs=True):
 
-        # if the coords do not align then set time dimensions equal
+        # --------------------------------------------------------------
+        # Align time coords if needed
+        # --------------------------------------------------------------
         try:
             xr.align(dev, ref, join='exact')
-        except BaseException:
-            ref.coords["time"] = dev.coords["time"]
+        except ValueError:
+            ref = ref.assign_coords(time=dev.coords["time"])
 
-        # Compute absolute and fractional differences
-        # Also make sure to preserve variable attributes
-        absdiffs = dev.copy()
-        fracdiffs = dev.copy()
-        for var in varlist:
-            absdiffs[var].values = dev[var].values - ref[var].values
-            fracdiffs[var].values = dev[var].values / ref[var].values
-            absdiffs[var].attrs = dev[var].attrs
-            fracdiffs[var].attrs = dev[var].attrs
+        # --------------------------------------------------------------
+        # Compute the absolute and fractional diffs
+        # Use .load() to force read the variable's data into memory,
+        # which should avoid lazy indexing issues.  The per-variable
+        # data size is managaeable so overall performance should be OK.
+        # --------------------------------------------------------------
+        absdiffs = xr.Dataset(
+            {var: dev[var].load() - ref[var].load() for var in varlist}
+        )
+        fracdiffs = xr.Dataset(
+            {var: dev[var].load() / ref[var].load() for var in varlist}
+        )
 
-        # If there are more than one time point, then take the
-        # time mean (e.g. Annual Mean) of differences
-        if "time" in absdiffs.coords:
-            if len(absdiffs.coords["time"]) > 1:
-                absdiffs = dataset_mean(absdiffs)
-        if "time" in fracdiffs.coords:
-            if len(fracdiffs.coords["time"]) > 1:
-                fracdiffs = dataset_mean(fracdiffs)
+        # --------------------------------------------------------------
+        # Take the time mean if there is more than one timestamp
+        # --------------------------------------------------------------
+        if "time" in absdiffs.coords and len(absdiffs.coords["time"]) > 1:
+            absdiffs = dataset_mean(absdiffs)
+        if "time" in fracdiffs.coords and len(fracdiffs.coords["time"]) > 1:
+            fracdiffs = dataset_mean(fracdiffs)
 
     return absdiffs, fracdiffs